+++ var/template/api/events/main-events.js
@@ -1,7 +1,0 @@
-module.exports = {
-    hello: function * (params) {
-        return {
-            response: 'Hello'
-        };
-    }
-};