+++ var/template/config/routes.js
@@ -1,5 +1,0 @@
-exports.routes = {
-    'GET /hello' : {
-        method: 'main.hello'
-    }
-};